srds augeas REFACTOR init and destroy reworked
We need to initialize augeas technically once
for every connection but the DS init and destroy
callbacks are called only when the module is
being installed/uninstalled. So we need to perform
init in every callback, if needed, and no place
to put destroy. Needs changes in DS plugin API
for proper support.