augyang BUGFIX in ay_ynode_delete_build_list_()
In the ay_lnode_lense_equal(), lenses are compared regardless of tag.
Thus, there is no difference between the tag 'label' and 'key', but
it depends on the string. Also, the ay_ynode_delete_build_list_()
has been modified so that it does not matter the node type when
detecting the 'lns . (sep . lns) . (sep . lns)*' pattern.
Fix based on yum.aug.