augyang BUGFIX merge cases only by value
When merging cases, it is necessary to distinguish the situation where
the branches differ only based on the value in the first node. In that
situation, the second branch can be deleted and the 'when' does not
have to be solved at all (and also merging as such). Additionally,
the dovecot.yang module was added.