augyang BUGFIX 'when' target cannot be in grouping
Actually, the 'when' target can be in the grouping, but it can make
the yang confusing and it also complicates the construction of the
'when' path a bit, because the grouping can contain other
uses-stmt, etc. Also the path construction has been fixed because
the target can be in list. Fix based od yum.aug.