augyang BUGFIX insert_case and choices
Fixed case-stmt generation to better recognize concatenations in lense. In certain cases the ay_ynode.choice value is
newly switched from the state 'unified' to 'origin'. This fixed the mandatory-stmt in the 'nginx' module. The 'when'
statement can be in comment if its value is not valid from the point of view of XPATH 1.0.
Fix based on shellvars_list.aug and module is added too.