augyang CHANGE list with '_seq' key
A 'seq_list' is a special type of list that contains the '_seq' node as
a key. This key is used as a counter in lenses and typically indicates
the order of records. Thanks to this commit, the unnecessary container
was deleted and the value of the augeas counter is moved as a key in
the list.