augyang BUGFIX mandatory-stmt in choice and list
Mandatory evaluation in list and choice has been overwritten.
New rules have been added that decide the mandatory status.
The general view is that mandatory-true is nice to have, because it
increases the readability of the YANG module and simplifies,
for example, data storage, testing, etc. But if mandatory-true is set,
but it's not supposed to be, then it's an error in the generated
YANG module.