augyang BUGFIX merge choice-cases from the bottom
The order in which nodes are merged has changed. Before, the merging
took place from the top-level layer, now it's from the bottom (leafs).
Thanks to this, the merging of the top nodes will be done properly.
Moreover, merging is done as long as possible. Also, deleting
non-unique branches in choice has been moved to a separate function.