augyang CHANGE create identifier with no_case flag
If the node contains a pattern that actually contains identifiers,
then the node is divided into several nodes and each of them is
assigned an identifier. The problem is deciding what identifiers
to create if case does not matter. The guess is that most often
all letters will be lowercase or all uppercase or only the first
one is uppercase. This commit implements the first two cases.
The third case may be implemented in the future.